MSVC fixes

+++ solvers/qpSpline/splineGeneration.cpp
@@ -103,7 +103,7 @@ PiecewisePolynomial<double> generateSpline(const SplineInformation& spline_infor
 }
 
 PiecewisePolynomial<double> nWaypointCubicSpline(const vector<double>& segment_times, double x0, double xd0, double xf, double xdf, const Ref<const VectorXd> &xi) {
-  const int num_segments = xi.size()+1;
+  const int num_segments = static_cast<int>(xi.size() + 1);
   assert(segment_times.size() == num_segments + 1);
 
   int polynomial_order = 3;
